+++ crackerjack/__init__.py
@@ -1,6 +1,5 @@
 import typing as t
 
-
 from acb import register_pkg
 
 register_pkg("crackerjack")